fix(#95): address CodeRabbit feedback — clamp typing + room-switch race
- computeLinkQuality: lossPctRaw.clamp() returns num; spell .toDouble()
  explicitly so the record's `double lossPct` field is satisfied without
  relying on the analyzer's flow-typing.

- _sendLinkQuality: re-check the room snapshot after each await via a
  new _stillSameGuestRoomFor helper. A leave + rejoin during the
  telemetry round-trip would otherwise re-seed _prevTelemetry for the
  old room's mac after teardown cleared it, and could send a
  LinkQuality on a transport whose peer was already forgotten.

(CodeRabbit also suggested bumping kProtocolVersion. Skipped: the
current decoder drops unknown kinds via FormatException — the desired
forward-compat behaviour. Bumping the envelope version would make every
v1 peer reject every v2 message at the version check, breaking even
unchanged kinds. The project's existing convention — see voicePsm
added to JoinAccepted in earlier work — is to keep v=1 for additive
kind additions.)

+++ lib/services/link_quality_reporter.dart
@@ -59,7 +59,10 @@ import 'audio_service.dart';
 
   final lossPctRaw =
       expectedFrames <= 0 ? 0.0 : (lateDelta / expectedFrames) * 100.0;
-  final lossPct = lossPctRaw.clamp(0.0, 100.0);
+  // `num.clamp` returns `num`, not `double`; the record's `lossPct` field is
+  // typed `double`, so spell the conversion explicitly to keep the analyzer
+  // (and a strict-mode reader) happy.
+  final lossPct = lossPctRaw.clamp(0.0, 100.0).toDouble();
 
   final jitterMs = current.currentDepthFrames * frameDurationMs;
   final underrunsPerSec = underrunDelta / (elapsedMs / 1000.0);
